move the logic of getPresignedUploadURL and getPresignedDownloadURL to /internal/file/file.go
